requirements.txt
# FASHN Human Parser - clothing segmentation
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0
numpy>=1.21.0

# Production parser (recommended for best accuracy)
fashn-human-parser>=0.1.0

# Optional: visualization
matplotlib>=3.5.0

# Web app
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
